voxir_app/requirements.txt
Flask>=2.0
pyttsx3>=2.90
Pillow>=9.0
pytesseract>=0.3.9
opencv-python>=4.5
moviepy>=1.0.3
